pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "homeassistant-mcp-server"
version = "1.0.0"
description = "Model Context Protocol server for Home Assistant integration"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Home Assistant MCP", email = "dev@example.com"}
]
license = {text = "MIT"}
dependencies = [
    "mcp>=1.0.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
]

[project.scripts]
homeassistant-mcp-server = "homeassistant_mcp_server.server:main"

[project.urls]
Homepage = "https://github.com/yourusername/homeassistant-mcp-server"
Repository = "https://github.com/yourusername/homeassistant-mcp-server"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"